refactor: list bullet width flexible
Its bad when its a fixed size. If text is too big all of a sudden you start getting the . in 1. on a seperate line.

This design was inherited from markdown_widget. I've been fine with it in markdown_widget but I like the result better *without* the fixed size, even setting aside wrapping issues.